refactor: split AnimationParser class
