+++ Editor/Processors/TraceAndOptimize/AnimationParser.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 7660a178396f4b939a47a87ceb191329
+timeCreated: 1692885437
