+++ Editor/Processors/TraceAndOptimize/AnimatorParser.cs
@@ -4,7 +4,6 @@ using System.Linq;
 using System.Reflection;
 using static Anatawa12.AvatarOptimizer.ErrorReporting.BuildReport;
 using JetBrains.Annotations;
-using UnityEditor;
 using UnityEditor.Animations;
 using UnityEngine;
 using UnityEngine.Animations;
@@ -18,6 +17,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
     {
         private bool mmdWorldCompatibility;
         private bool advancedAnimatorParser;
+        private AnimationParser _animationParser = new AnimationParser();
 
         public AnimatorParser(bool mmdWorldCompatibility, bool advancedAnimatorParser)
         {
@@ -79,7 +79,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
                 // so if `Play Automatically` is disabled, we can do nothing with Animation component.
                 // That's why we ignore Animation component if playAutomatically is false.
 
-                var parsed = GetParsedAnimation(gameObject, animation.clip);
+                var parsed = _animationParser.GetParsedAnimation(gameObject, animation.clip);
 
                 var alwaysApplied =
                     objectAlwaysActive
@@ -508,7 +508,7 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
         /// </summary>
         private IModificationsContainer FallbackParseAnimatorController(GameObject root, RuntimeAnimatorController controller)
         {
-            return controller.animationClips.Select(clip => GetParsedAnimation(root, clip)).MergeContainersSideBySide();
+            return controller.animationClips.Select(clip => _animationParser.GetParsedAnimation(root, clip)).MergeContainersSideBySide();
         }
 
         internal IModificationsContainer AdvancedParseAnimatorController(GameObject root, AnimatorController controller,
@@ -563,61 +563,15 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
             if (syncedLayer == -1)
             {
                 parsedMotions = CollectStates(layer.stateMachine)
-                    .Select(state => ParseMotion(root, state.motion, mapping));
+                    .Select(state => _animationParser.ParseMotion(root, state.motion, mapping));
             }
             else
             {
                 parsedMotions = CollectStates(controller.layers[syncedLayer].stateMachine)
-                    .Select(state => ParseMotion(root, layer.GetOverrideMotion(state), mapping));
+                    .Select(state => _animationParser.ParseMotion(root, layer.GetOverrideMotion(state), mapping));
             }
 
             return parsedMotions.MergeContainersSideBySide();
-        }
-
-        internal IModificationsContainer ParseMotion(GameObject root, Motion motion,
-            IReadOnlyDictionary<AnimationClip, AnimationClip> mapping) =>
-            ReportingObject(motion, () => ParseMotionInner(root, motion, mapping));
-
-        private IModificationsContainer ParseMotionInner(GameObject root, Motion motion,
-            IReadOnlyDictionary<AnimationClip, AnimationClip> mapping)
-        {
-            switch (motion)
-            {
-                case null:
-                    return ImmutableModificationsContainer.Empty;
-                case AnimationClip clip:
-                    return GetParsedAnimation(root, mapping.TryGetValue(clip, out var newClip) ? newClip : clip);
-                case BlendTree blendTree:
-                    return ParseBlendTree(root, blendTree, mapping);
-                default:
-                    LogFatal("Unknown Motion Type: {0} in motion {1}", motion.GetType().Name, motion.name);
-                    return ImmutableModificationsContainer.Empty;
-            }
-        }
-
-        private IModificationsContainer ParseBlendTree(GameObject root, BlendTree blendTree,
-            IReadOnlyDictionary<AnimationClip, AnimationClip> mapping)
-        {
-            switch (blendTree.blendType)
-            {
-                case BlendTreeType.Simple1D:
-                case BlendTreeType.SimpleDirectional2D:
-                case BlendTreeType.FreeformDirectional2D:
-                case BlendTreeType.FreeformCartesian2D:
-                    // in those blend blend total blend is always 1 so
-                    // if all animation sets same value, the result will also be same value.
-                    return blendTree.children.Select(x => ParseMotionInner(root, x.motion, mapping)).MergeContainersSideBySide();
-                case BlendTreeType.Direct:
-                    // in direct blend tree, total blend can be not zero so all properties are Variable.
-                    var merged = blendTree.children.Select(x => ParseMotionInner(root, x.motion, mapping))
-                        .MergeContainersSideBySide()
-                        .ToMutable();
-
-                    merged.MakeAllVariable();
-                    return merged;
-                default:
-                    throw new ArgumentOutOfRangeException();
-            }
         }
 
         private IEnumerable<AnimatorState> CollectStates(AnimatorStateMachine stateMachineIn)
@@ -664,45 +618,6 @@ namespace Anatawa12.AvatarOptimizer.Processors.TraceAndOptimizes
                     if (!overrides.ContainsKey(original))
                         overrides.Add(original, mapped);
             }
-        }
-
-        #endregion
-
-        #region Animaton
-
-        private readonly Dictionary<(GameObject, AnimationClip), ImmutableModificationsContainer> _parsedAnimationCache =
-            new Dictionary<(GameObject, AnimationClip), ImmutableModificationsContainer>();
-
-        internal ImmutableModificationsContainer GetParsedAnimation(GameObject root, [CanBeNull] AnimationClip clip)
-        {
-            if (clip == null) return ImmutableModificationsContainer.Empty;
-            if (!_parsedAnimationCache.TryGetValue((root, clip), out var parsed))
-                _parsedAnimationCache.Add((root, clip), parsed = ParseAnimation(root, clip));
-            return parsed;
-        }
-
-        public static ImmutableModificationsContainer ParseAnimation(GameObject root, [NotNull] AnimationClip clip)
-        {
-            var modifications = new ModificationsContainer();
-
-            foreach (var binding in AnimationUtility.GetCurveBindings(clip))
-            {
-                var obj = AnimationUtility.GetAnimatedObject(root, binding);
-                if (obj == null) continue;
-                var componentOrGameObject = obj is Component component ? (ComponentOrGameObject)component
-                    : obj is GameObject gameObject ? (ComponentOrGameObject)gameObject
-                    : throw new InvalidOperationException($"unexpected animated object: {obj} ({obj.GetType().Name}");
-
-                var curve = AnimationUtility.GetEditorCurve(clip, binding);
-                var currentPropertyMayNull = AnimationProperty.ParseProperty(curve);
-
-                if (!(currentPropertyMayNull is AnimationProperty currentProperty)) continue;
-
-                modifications.ModifyObject(componentOrGameObject)
-                    .AddModificationAsNewLayer(binding.propertyName, currentProperty);
-            }
-
-            return modifications.ToImmutable();
         }
 
         #endregion
